Django/Trendz/requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
colorama==0.4.3
Django==3.0.5
django-tastypie==0.14.3
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.11.2
